requirements.txt
exifread
python-docx
PyPDF2
